# Inline `IsolatedExecutionState#state`